+++ CHANGES.md
@@ -9,7 +9,7 @@
 * Add the ability to route basted on the `dest` request header when using the
   TTwitter Thrift protocol
 * Change the default `dstPrefix` from the protocol name to `/s`
-* Change the default HTTP identifier to simply use the `Host` header
+* Change the default HTTP identifier to the `io.l5d.header.token` identifier
 
 ## 0.8.6 2017-01-19